refactor: moved permalink update to separate controller
