+++ subscriptions_holder/src/main/java/ru/dankoy/subscriptionsholder/subscriptions_holder/core/controller/CommunitySubController.java
@@ -1,6 +1,5 @@
 package ru.dankoy.subscriptionsholder.subscriptions_holder.core.controller;
 
-
 import jakarta.validation.Valid;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -10,25 +9,19 @@ import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
-import ru.dankoy.subscriptionsholder.subscriptions_holder.core.dto.SubscriptionUpdatePermalinkDTO;
 import ru.dankoy.subscriptionsholder.subscriptions_holder.core.dto.communitysub.CommunitySubCreateDTO;
 import ru.dankoy.subscriptionsholder.subscriptions_holder.core.dto.communitysub.CommunitySubDTO;
 import ru.dankoy.subscriptionsholder.subscriptions_holder.core.service.CommunitySubService;
-import ru.dankoy.subscriptionsholder.subscriptions_holder.core.service.SubscriptionService;
-
 
 @RequiredArgsConstructor
 @RestController
 public class CommunitySubController {
 
   private final CommunitySubService communitySubService;
-  private final SubscriptionService subscriptionService;
-
 
   @GetMapping(value = "/api/v1/community_subscriptions")
   public Page<CommunitySubDTO> getAll(Pageable pageable) {
@@ -36,25 +29,25 @@ public class CommunitySubController {
     return s.map(CommunitySubDTO::toDTO);
   }
 
-
-  @GetMapping(value = "/api/v1/community_subscriptions", params = {"active"})
+  @GetMapping(
+      value = "/api/v1/community_subscriptions",
+      params = {"active"})
   public Page<CommunitySubDTO> getAllWithActiveChats(
       @RequestParam(value = "active", defaultValue = "true") boolean active, Pageable pageable) {
 
     var s = communitySubService.findAllByChatActive(active, pageable);
 
     return s.map(CommunitySubDTO::toDTO);
-
   }
 
-  @GetMapping(value = "/api/v1/community_subscriptions", params = {"telegramChatId"})
+  @GetMapping(
+      value = "/api/v1/community_subscriptions",
+      params = {"telegramChatId"})
   public List<CommunitySubDTO> getAllByTelegramChatId(
       @RequestParam(value = "telegramChatId") long telegramChatId) {
     var s = communitySubService.getAllByChatChatId(telegramChatId);
     return s.stream().map(CommunitySubDTO::toDTO).toList();
-
   }
-
 
   @PostMapping(path = "/api/v1/community_subscriptions")
   public CommunitySubDTO subscribeChatToCommunity(@Valid @RequestBody CommunitySubCreateDTO dto) {
@@ -64,9 +57,7 @@ public class CommunitySubController {
     var s = communitySubService.subscribeChatToCommunity(subscription);
 
     return CommunitySubDTO.toDTO(s);
-
   }
-
 
   @DeleteMapping(path = "/api/v1/community_subscriptions")
   @ResponseStatus(code = HttpStatus.ACCEPTED)
@@ -75,19 +66,5 @@ public class CommunitySubController {
     var subscription = CommunitySubCreateDTO.fromDTO(dto);
 
     communitySubService.unsubscribeChatFromCommunity(subscription);
-
   }
-
-  @PutMapping(path = "/api/v1/community_subscriptions")
-  public SubscriptionUpdatePermalinkDTO updatePermalink(
-      @Valid @RequestBody SubscriptionUpdatePermalinkDTO dto) {
-
-    var subscription = SubscriptionUpdatePermalinkDTO.fromDTO(dto);
-
-    var s = subscriptionService.updatePermalink(subscription);
-
-    return SubscriptionUpdatePermalinkDTO.toDTO(s);
-
-  }
-
 }
